Add files via upload
Matlab code for solving the two dimensional wave equation with constant values at the quadrangular domain (Dirichlet boundary conditions for x and y) and initial derivative in time equals given, for all x and y inside the domain. In addition, the profile u(x,y,0) must be informed bu the user in the first plot zone of the code. The code also saves a video animation for the plot as time evolves, in order to see what' happening after all iterations.